tests/requirements.txt
pytest
python-dotenv
pymongo[srv]
supabase
psycopg2-binary
sqlalchemy
